requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
boto3==1.34.24
newrelic>=11.0.0,<12.0.0
requests==2.31.0
